changes to model, bug fixes,performance enhancements and addtion of valve change limition

+++ lib/app/data/models/all_devices_model.dart
@@ -81,8 +81,8 @@ class IoTDevices {
 class Readings {
   int? id;
   int? deviceId;
-  int? litersConsumed;
-  int? flowRate;
+  double? litersConsumed;
+  double? flowRate;
   String? createdAt;
   String? updatedAt;
